+++ app/dashboard/layout.tsx
@@ -1,7 +1,8 @@
 import { PropsWithChildren } from 'react';
 
+import DashboardNav from '@components/organisms/DashboardNav';
+
 import Page from './_components/atoms/Page';
-import DashboardNav from '../../components/organisms/DashboardNav';
 
 type DashboardLayoutProps = PropsWithChildren;
